ci: Add GitHub Actions workflow for running UI tests
This commit adds a GitHub Actions workflow for running UI tests on Linux, Windows, and Mac. The workflow prepares and launches IntelliJ IDEA in UI testing mode which is needed to interact with UI and runs UI tests using a Gradle task. Three jobs are defined in the workflow - one for each operating system. The `Health Check` step waits for IDEA to start and the `Tests` step runs the tests. The commit also adds the Run IDE for UI Tests project configuration.

Signed-off-by: Milad Nekofar <milad@nekofar.com>