refactor: use provider for properties instead of deprecated API
The properties implementation is updated to use provider API instead of the deprecated API. This commit modifies the deprecated methods to avoid using them. There’s a code change in lines 5, 6, 17, and 18 setting the values of group and version properties. Also, there's a change in line 19 in retrieving the environment variables to apply proper methods.

fixup! refactor: use provider for properties instead of deprecated API

fixup! refactor: use provider for properties instead of deprecated API