ci: remove GitHub Action for Qodana viral analysis
This commit removes a GitHub workflow that performed viral analysis on the code using Qodana. The workflow was triggered on pushes to the 'hotfix/*' and 'release/*' branches and pull requests that were opened or reopened or were manually activated. It used the Qodana Action to run the Qodana Code Inspection tool and upload the report in SARIF format using the CodeQL Action. The Concurrency group of the workflow has also been removed.