ci: replace manual GitHub release creation with action
Refactoring the 'Create Release Draft' section of the GitHub Actions workflow to use the action 'nekofar/create-github-release@v1.0.8'. This switch simplifies the release creation by removing the need for manual scripting, making the GitHub Actions workflow cleaner and more efficient. It also eases the process of marking the release as a pre-release under certain conditions.